% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjmat_truth}
\alias{adjmat_truth}
\title{Baseline dataset (precision matrix) used for simulating X_n500.rds and
Y_n500.rda}
\format{
An object of class \code{data.frame} with 15 rows and 15 columns.
}
\usage{
adjmat_truth
}
\description{
Baseline dataset (precision matrix) used for simulating X_n500.rds and
Y_n500.rda
}
\keyword{datasets}
