% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B_Rho_Phi_alphap_alphaq.R
\name{B_Rho_Phi_alphap_alphaq}
\alias{B_Rho_Phi_alphap_alphaq}
\title{Internal function}
\usage{
B_Rho_Phi_alphap_alphaq(input_list)
}
\arguments{
\item{input_list}{...}
}
\description{
Internal function
}
