% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_transform_Phi.R
\name{baseline_transform_Phi}
\alias{baseline_transform_Phi}
\title{Perform baseline transformation on Phi}
\usage{
baseline_transform_Phi(Phi, Y_levels)
}
\arguments{
\item{Phi}{...}

\item{Y_levels}{...}
}
\description{
Perform baseline transformation on Phi
}
