% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_transform_Rho.R
\name{baseline_transform_Rho}
\alias{baseline_transform_Rho}
\title{Perform baseline transformation on Rho}
\usage{
baseline_transform_Rho(Rho, Y_levels)
}
\arguments{
\item{Rho}{...}

\item{Y_levels}{...}
}
\description{
Perform baseline transformation on Rho
}
