% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_function.R
\name{BIC_function}
\alias{BIC_function}
\title{Internal function}
\usage{
BIC_function(x, X, Y, Y_levels, groups_X, prior_X, use_prior)
}
\arguments{
\item{x}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}

\item{groups_X}{...}

\item{prior_X}{...}

\item{use_prior}{...}
}
\description{
Internal function
}
