% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MGM.R
\name{fit_MGM}
\alias{fit_MGM}
\title{Calculate a mixed graphical model with or without additional priors (PriOmics algorithm)}
\usage{
fit_MGM(
  X,
  Y,
  groups_X = FALSE,
  lambda_seq,
  iterations = 100,
  eps = 1e-06,
  prior_X = FALSE
)
}
\arguments{
\item{X}{continuous data (n x p matrix or data.frame)}

\item{Y}{discrete/categorical data (n x p matrix or data.frame)}

\item{groups_X}{list of groups for prior assumption (for continuous data)}

\item{lambda_seq}{list of tuning parameter lambda (model selection)}

\item{iterations}{maximum number of iterations, until optimization is stopped}

\item{eps}{precision parameter}

\item{prior_X}{vector of prior assumption ("A", "B" or "A_exp" (not recommended)) for each group specified in "groups_X"}
}
\value{
list of resulting models according to lambda value
}
\description{
Calculate a mixed graphical model with or without additional priors (PriOmics algorithm)
}
\examples{
\dontrun{
################################################################################
### fit MGM without prior assumptions
################################################################################
# load sample data
data(X_n500)
data(Y_n500)
data(adjmat_truth)

# 1a) fit MGM
fit1 <- fit_MGM(X = X_n500,
                Y = Y_n500,
                lambda_seq = c(1.3^c(-2:-17)),
                iterations = 500
)

# 1b) fit MGM in parallel mode
fit1 <- fit_MGM_par(X = X_n500,
                    Y = Y_n500,
                    lambda_seq = c(1.3^c(-2:-17)),
                    iterations = 500,
                    n_cor = 6
)

################################################################################
# 2) Model selection:
# inspection of model parameter (edges, iterations, BIC, EBIC)
plot_BIC(fit1)
# manually select 9th lambda value in sequence and generate adjacency matrix
adj <- get_adj_mat(fit1, 9)
# or automatically
adj <- get_adj_mat(fit1, mgm_stats(fit1)[[2]]) # EBIC/BIC optimum

################################################################################
# 3) plot heatmap of adjacency matrix and compare it to original data matrix
plot_heat(adj)
plot_heat(adjmat_truth)
plot_heat_selection(adj, "D|F|X_1") # plot subset of adj

################################################################################
# 4) interactive Plots
# a) interactive networks
plot_net(adj, plot_style = "circle")
plot_net(adj, plot_style = "igraph")

# save network as html
net <- plot_net(adj, plot_style = "circle")
visNetwork::visSave(net, file = ".../path/.../net.html")

# b) interactive heatmaps (for larger matrices)
plot_iheat(adj)
# save heatmap as html
plot_iheat(adj, save_file = T)


################################################################################
### fit MGM with prior A (redundant feature groups, i.e. copies)
################################################################################

# add known structure to model
X_groups <- list(1, 2, c(3, 4), c(5, 6), c(7, 8, 9), c(10, 11, 12))
prior <- rep("A", length(X_groups))

fit2 <- fit_MGM(X = X_n500,
                Y = Y_n500,
                groups_X = X_groups,
                prior_X = prior,
                lambda_seq = c(1.3^c(-2:-17)),
                iterations = 500

)

plot_BIC(fit2)
adj2 <- get_adj_mat(fit2, 6)

plot_heat(adj2)
plot_heat(adjmat_truth)

################################################################################
### fit MGM with prior B (similar feature groups, but no copies)
################################################################################

# add known structure to model
X_groups <- list(1, 2, c(3, 4), c(5, 6), c(7, 8, 9), c(10, 11, 12))
prior <- rep("B", length(X_groups))

fit3 <- fit_MGM(X = X_n500,
                Y = Y_n500,
                groups_X = X_groups,
                prior_X = prior,
                lambda_seq = c(1.3^c(-2:-17)),
                iterations = 500
)

plot_BIC(fit3)
adj3 <- get_adj_mat(fit3, 9)

plot_heat(adj3)
plot_heat(adjmat_truth)
}

}
