% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_MGM_par.R
\name{fit_MGM_par}
\alias{fit_MGM_par}
\title{Parallel calculation of a mixed graphical model with or without additional
priors (PriOmics algorithm)}
\usage{
fit_MGM_par(
  X,
  Y,
  groups_X = FALSE,
  lambda_seq,
  iterations = 100,
  eps = 1e-06,
  prior_X = FALSE,
  n_cor = 6
)
}
\arguments{
\item{X}{continuous data (n x p matrix or data.frame)}

\item{Y}{discrete/categorical data (n x p matrix or data.frame)}

\item{groups_X}{list of groups for prior assumption (for continuous data)}

\item{lambda_seq}{list of tuning parameter lambda (model selection)}

\item{iterations}{maximum number of iterations, until optimization is stopped}

\item{eps}{precision parameter}

\item{prior_X}{vector of prior assumption ("A", "B" or "A_exp" (not recommended)) for each group specified in "groups_X"}

\item{n_cor}{number of CPU cores}
}
\value{
list of resulting models according to lambda value
}
\description{
Parallel calculation of a mixed graphical model with or without additional
priors (PriOmics algorithm)
}
