% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adj_mat.R
\name{get_adj_mat}
\alias{get_adj_mat}
\title{Combine cont-cont (B), cont-disc (Rho) & disc-disc (Phi) couplings to an
adjacency matrix for a specified fit}
\usage{
get_adj_mat(fit, lambda)
}
\arguments{
\item{fit}{fit calculated in fit_MGM() or fit_MGM_par()}

\item{lambda}{select a lambda id calculated in fit}
}
\description{
Combine cont-cont (B), cont-disc (Rho) & disc-disc (Phi) couplings to an
adjacency matrix for a specified fit
}
