% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adj_mat2.R
\name{get_adj_mat2}
\alias{get_adj_mat2}
\title{Internal function}
\usage{
get_adj_mat2(B, Rho, Phi)
}
\arguments{
\item{B}{continuous-continuous couplings (B)}

\item{Rho}{continuous-discrete couplings (Rho)}

\item{Phi}{discrete-discrete couplings (Phi)}
}
\description{
Internal function
}
