% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_IDs_and_weights.R
\name{get_group_IDs_and_weights}
\alias{get_group_IDs_and_weights}
\title{Internal function}
\usage{
get_group_IDs_and_weights(IDs_B, IDs_rho, X, Y, Y_levels)
}
\arguments{
\item{IDs_B}{...}

\item{IDs_rho}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}
}
\description{
Internal function
}
