% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Extract group information from column names in continuous data (i.e., X).
Group names should be indicated by an underscore, e.g. Group_Feature or
Protein_Peptide.}
\usage{
get_groups(x)
}
\arguments{
\item{x}{matrix or data.frame with continuous data}
}
\description{
Extract group information from column names in continuous data (i.e., X).
Group names should be indicated by an underscore, e.g. Group_Feature or
Protein_Peptide.
}
