% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_f_temp.R
\name{grad_f_temp}
\alias{grad_f_temp}
\title{Internal function}
\usage{
grad_f_temp(x, X, Y, Y_levels, p, q, groups_X, prior_X, use_prior)
}
\arguments{
\item{x}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}

\item{p}{...}

\item{q}{...}

\item{groups_X}{...}

\item{prior_X}{...}

\item{use_prior}{...}
}
\description{
Internal function
}
