% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_neglogli_plain.R
\name{grad_neglogli_plain}
\alias{grad_neglogli_plain}
\title{Internal function}
\usage{
grad_neglogli_plain(
  B_Rho_Phi_alphap_alphaq,
  X,
  Y,
  Y_levels,
  p,
  q,
  groups_X,
  prior_X,
  use_prior
)
}
\arguments{
\item{B_Rho_Phi_alphap_alphaq}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}

\item{p}{...}

\item{q}{...}

\item{groups_X}{...}

\item{prior_X}{...}

\item{use_prior}{...}
}
\description{
Internal function
}
