% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_neglogli_R.R
\name{grad_neglogli_R}
\alias{grad_neglogli_R}
\title{Internal function}
\usage{
grad_neglogli_R(
  list_Inv_B_Rho_Phi_alphap_alphaq,
  X,
  Y,
  Y_levels,
  groups_X,
  prior_X,
  use_prior
)
}
\arguments{
\item{list_Inv_B_Rho_Phi_alphap_alphaq}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}

\item{groups_X}{...}

\item{prior_X}{...}

\item{use_prior}{...}
}
\description{
Internal function
}
