% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inv_B_Rho_Phi_alphap_alphaq.R
\name{Inv_B_Rho_Phi_alphap_alphaq}
\alias{Inv_B_Rho_Phi_alphap_alphaq}
\title{Internal function}
\usage{
Inv_B_Rho_Phi_alphap_alphaq(x, p, q)
}
\arguments{
\item{x}{...}

\item{p}{...}

\item{q}{...}
}
\description{
Internal function
}
