% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_starting_parameters.R
\name{make_starting_parameters}
\alias{make_starting_parameters}
\title{Internal function}
\usage{
make_starting_parameters(X, Y)
}
\arguments{
\item{X}{...}

\item{Y}{...}
}
\description{
Internal function
}
