% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgm_stats.R
\name{mgm_stats}
\alias{mgm_stats}
\title{Get information about model fits}
\usage{
mgm_stats(fit)
}
\arguments{
\item{fit}{calculated with fit_MGM() or fit_MGM_par()}
}
\description{
Get information about model fits
}
