% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neglogli.R
\name{neglogli}
\alias{neglogli}
\title{Internal function}
\usage{
neglogli(input_list, X, Y, Y_levels, groups_X, prior_X, use_prior)
}
\arguments{
\item{input_list}{...}

\item{X}{...}

\item{Y}{...}

\item{Y_levels}{...}

\item{groups_X}{...}

\item{prior_X}{...}

\item{use_prior}{...}
}
\description{
Internal function
}
