% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_BIC.R
\name{plot_BIC}
\alias{plot_BIC}
\title{Plot overview of PriOmics models (Lambda against BIC/EBIC)}
\usage{
plot_BIC(
  x,
  mar = c(10, 7, 4, 7),
  cex.bic.text = 0.7,
  cex.it.text = 0.8,
  leg.position = "top",
  legend.text.size = 0.7,
  cex.axis = 1.5,
  cex.lambda.text = 1.5,
  cex.bic.axis = 1.5,
  cex.ebic.axis = 1.5
)
}
\arguments{
\item{x}{list of fits calculated with fit_MGM() or fit_MGM_par()}
}
\description{
Plot overview of PriOmics models (Lambda against BIC/EBIC)
}
