% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat.R
\name{plot_heat}
\alias{plot_heat}
\title{Plot basic heatmap of adjacency matrix (based on package 'pheatmap')}
\usage{
plot_heat(x, covtocor = FALSE)
}
\arguments{
\item{x}{result from get_adj_mat()}
}
\description{
Only useful for limited number of features (p < 50)
}
