% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iheat.R
\name{plot_iheat}
\alias{plot_iheat}
\title{Plot interactive heatmap of adjacency matrix (based on package'heatmaply')}
\usage{
plot_iheat(x, upper_only = FALSE, dark_colors = FALSE, save_file = FALSE)
}
\arguments{
\item{x}{result from get_adj_mat()}

\item{upper_only}{logical, If TRUE, only the upper half of the matrix is
plotted. (default = FALSE)}

\item{dark_colors}{logical, If TRUE, a darker color palette
is used.  (default = FALSE)}

\item{save_file}{logical, If TRUE, the heatmap is saved as html file at
current working directory (default = FALSE); no visible output
Saving (and opening) files, may take a while:
p=100 -> ~5mb (5s)
p=500 -> ~25mb (30s)
p=1000 -> ~90mb (2min)
p=2000 -> ~330mb (7min) not recommended! (R crashes, but file will be stored
anyway)}
}
\description{
For large matrices zoom into the plot
}
