% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net.R
\name{plot_net}
\alias{plot_net}
\title{Plot network based on package 'visNetwork'}
\usage{
plot_net(x, plot_style, node_color, node_shape, node_scale, node_bold)
}
\arguments{
\item{x}{adjacency matrix (generated by get_adj_mat())}

\item{plot_style}{"circle", "igraph" or "visNetwork" (useful only for few features)}

\item{node_color}{vector of colors for nodes (default is rainbow() color set)}

\item{node_shape}{vector of shapes for nodes (e.g., "dot", "rectangle", "triangle",
"hexagon", "diamond", "ellipse", "star")}

\item{node_scale}{logical, If TRUE, node sizes are proportional to feature
impact (default = FALSE)}

\item{node_bold}{logical, If TRUE, node titles are displayed as bold font
(default = FALSE)}
}
\description{
Positive edges are displayed in blue, negative edges in red. The color
depth represents the association strength.
}
