% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_old.R
\name{plot_net_old}
\alias{plot_net_old}
\title{DEPRECATED: Plot network based on package 'visNetwork'}
\usage{
plot_net_old(x, plot_style, node_color, node_shape)
}
\arguments{
\item{x}{adjacency matrix (generated by get_adj_mat())}

\item{plot_style}{"circle", "igraph" or "visNetwork" (useful only for few features)}

\item{node_color}{vector of colors for nodes (default is rainbow() color set)}

\item{node_shape}{vector of shapes for nodes (e.g., "dot", "rectangle", "triangle",
"hexagon", "diamond", "ellipse", "star")}
}
\description{
Positive edges are displayed in blue, negative edges in red. The color
depth represents the association strength.
}
