% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_net_selection.R
\name{plot_net_selection}
\alias{plot_net_selection}
\title{Plot first order neighborhood network based on package 'visNetwork'}
\usage{
plot_net_selection(adj, x, plot_style, node_color, node_shape)
}
\arguments{
\item{adj}{adjacency matrix (generated by get_adj_mat())}

\item{x}{target node for first order neighborhood}

\item{plot_style}{"circle" (default), "igraph" or "visNetwork" (useful only for few features)}

\item{node_color}{vector of node color(s) (default is rainbow() color set)}

\item{node_shape}{vector of shapes for nodes (e.g., "dot" (default), "rectangle", "triangle",
"hexagon", "diamond", "ellipse", "star")}
}
\description{
Positive edges are displayed in blue, negative edges in red. The color
depth represents the association strength.
}
