% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_grouplasso.R
\name{prox_grouplasso}
\alias{prox_grouplasso}
\title{Internal function for proximal operator of group lasso penalty}
\usage{
prox_grouplasso(x, t, opts = list(groups = as.list(seq(length(xc)))))
}
\arguments{
\item{x}{...}

\item{t}{...}

\item{opts}{...}
}
\description{
Internal function for proximal operator of group lasso penalty
}
\references{
adapted from github.com/jpvert/apg
}
