% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_mgm.R
\name{solve_mgm}
\alias{solve_mgm}
\title{Internal function}
\usage{
solve_mgm(
  grad_f,
  prox_g,
  xtemp,
  groupIDs = group_IDs,
  max_iters = 2,
  eps = 1e-06,
  alpha = 1.01,
  beta = 0.5
)
}
\arguments{
\item{grad_f}{...}

\item{prox_g}{...}

\item{xtemp}{...}

\item{groupIDs}{...}

\item{max_iters}{...}

\item{eps}{...}

\item{alpha}{...}

\item{beta}{...}
}
\description{
Internal function
}
