% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_B.R
\name{transform_B}
\alias{transform_B}
\title{Internal function}
\usage{
transform_B(B, groups_X, prior_X)
}
\arguments{
\item{B}{...}

\item{groups_X}{...}

\item{prior_X}{...}
}
\description{
Internal function
}
