% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_X.R
\name{transform_X}
\alias{transform_X}
\title{Internal function}
\usage{
transform_X(X, groups_X, prior_X)
}
\arguments{
\item{X}{...}

\item{groups_X}{...}

\item{prior_X}{...}
}
\description{
Internal function
}
