% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_n500}
\alias{X_n500}
\title{Simulated continuous data}
\format{
A data frame with 2000 rows and 12 variables:
}
\usage{
X_n500
}
\description{
A dataset containing artificial continuous data of 500 samples and 12
variables. The variables are either independent (A, B), duplicated (C, D) or
tripled (E, F), each with additional noise so simulate copies of the same
variable
}
\details{
The group structure looks as follows:
Group1 - A
Group2 - B
Group3 - C1, C2
Group4 - D1, D2
Group5 - E1, E2, E3
Group6 - F1, F2, F3
}
\keyword{datasets}
