% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Y_n500}
\alias{Y_n500}
\title{Simulated discrete data}
\format{
A data frame with 2000 rows and 3 variables:
}
\usage{
Y_n500
}
\description{
A dataset containing artificial discrete data of 500 samples and 3
variables.
}
\keyword{datasets}
